agentic-misra-do178c/pyproject.toml
[project]
name = "agentic-misra-do178c"
version = "0.1.0"
description = "Agentic AI code generation for MISRA-C + DO-178C (LangGraph/LCEL)"
authors = [{name="you"}]
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.2.22",
    "langchain-community>=0.2.10",
    "langgraph>=0.2.15",
    "pydantic>=2.5",
    "numpy>=1.26",
    "regex>=2024.4.16",
    "pdfminer.six>=20221105",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.7.0",
    "rank-bm25>=0.2.2",
    "pyyaml>=6.0.1",
    "rich>=13.7.1",
]

[tool.ruff]
line-length = 100
